protection - kamangir/bolt#746

+++ articraft/.abcli/ComfyUI/uninstall.sh
@@ -9,6 +9,11 @@ function ComfyUI_uninstall() {
         return
     fi
 
+    if [[ "$abcli_is_sagemaker_system" == true ]]; then
+        abcli_log_warning "image terminal command."
+        return 1
+    fi
+
     rm -rfv $abcli_path_git/ComfyUI
 
     abcli_conda rm ComfyUI
